test: 단위(H2)와 통합(Testcontainers) 테스트 분리 구성\n\n- add: H2 테스트 의존성 추가 및 ddl-auto create-drop 설정\n- add: test용 application.yml, tc 프로필용 application-tc.yml\n- add: Gradle integrationTest 태스크와 @Tag("integration") 분리\n- refactor: StrategicCitySimulatorApplicationTests 단순화 (Testcontainers 제거)\n- fix: H2 예약어 충돌 피하도록 Indicator.value -> metric_value 컬럼명 변경
