+++ src/test/java/com/example/repository/RepositoryTests.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@DataJpaTest
+@DataJpaTest(properties = "spring.jpa.hibernate.ddl-auto=create-drop")
 class RepositoryTests {
     @Autowired
     private GameSessionRepository gameSessionRepository;
